Image_captioning/requirements.txt
blis==0.4.1
catalogue==1.0.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.4
dataclasses==0.6
Flask==1.1.2
future==0.18.2
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
murmurhash==1.0.4
numpy==1.19.4
Pillow==8.0.1
plac==1.1.3
preshed==3.0.4
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
spacy==2.3.2
srsly==1.0.4
thinc==7.4.1
threadpoolctl==2.1.0
tqdm==4.53.0
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.0
Werkzeug==1.0.1
zipp==3.4.0